:recycle: テストのファイル分離をやめる
