+++ src/config.rs
@@ -1,4 +1,1 @@
 pub(crate) mod settings;
-
-#[cfg(test)]
-mod settings_tests;
